Added destination check

+++ victoria-clone.sh
@@ -12,6 +12,24 @@ debug=false
 size1=$((64 * 1024 * 1024 / 512))
 total_size=$((1280 * 1024 * 1024 / 512))
 
+# Functions
+function log() {
+	if [ $# -gt 1 ]; then
+		color=${1}
+		shift
+	else
+		color="123"
+	fi
+	echo -e "\033[38;5;${color}m${@}\033[m"
+}
+
+# Check for target disk
+chk=$(cat /proc/partitions | grep -m 1 ${dst_dev})
+if [ -z "${chk}" ]; then
+	log 204 "Could not find the destination device ${dst_dev}."
+	exit
+fi
+
 sfd0=$(sfdisk -d /dev/${src_dev})
 
 echo "size1 = ${size1}   total_size = ${total_size}"
@@ -19,7 +37,7 @@ echo ""
 echo -e "\033[38;5;211mSource partition scheme\033[m"
 echo "${sfd0}"
 
-sfd1=$(echo "${sfd0}" | sed -e "\/dev\/${src_base}2/s/start=[^,]*/start= P2_START/" -e "\/dev\/${src_base}2/s/size=[^,]*/start= P2_SIZE/")
+sfd1=$(echo "${sfd0}" | sed -e "\/dev\/${src_base}2/s/start=[^,]*/start= P2_START/" -e "\/dev\/${src_base}2/s/size=[^,]*/size= P2_SIZE/")
 
 if [ "${debug}" == "true" ]; then
 	echo ""
@@ -60,34 +78,29 @@ total_size=$((total * 512 / 1024 / 1024))
 
 echo ""
 echo "Last block = $((total - 1))  Total size = ${total_size} MB"
-echo ""
+
+echo "Proceed ? [y/n]"
+read resp
+resp=$(echo "${resp}" | tr '[:upper:]' '[:lower:]')
+if [ "${resp}" != "y" ] && [ "${resp}" != "yes" ]; then
+	exit
+fi
 
 # ===
 
-exit
-
-function log() {
-	echo -e "\033[38;5;123m${@}\033[m"
-}
-
-echo
+echo ""
 log "Using dd for the first 4 MB ..."
+
 
 dd if=/dev/${src_dev} of=/dev/${dst_dev} bs=1M count=4
 
-echo
-log "Applying new parition scheme ..."
+echo ""
+log "Applying new partition scheme ..."
 
 sfdisk --force /dev/${dst_base} > /dev/null <<< "${sfd1}"
 sync
 
-#sleep 1
-
-# partprobe /dev/${dst_dev}
-
-# sleep 1
-
-yes | mkfs -t vfat -F 32 /dev/${dst_base}1
+yes | mkfs -t vfat -F 32 -n ${label} /dev/${dst_base}1
 yes | mkfs -t ext4 -L ${label} /dev/${dst_base}2
 
 sync
@@ -99,9 +112,17 @@ rsync_options="--force -rltWDEHXAgoptx"
 rm -rf /mnt/clone
 mkdir -p /mnt/clone
 
-mount /dev/${dst_base}2 /mnt/clone
-sync
-rsync ${rsync_options} --delete --exclude '.gvfs' --exclude '/dev/*' --exclude '/mnt/clone/*' --exclude '/proc/*' --exclude '/run/*' --exclude '/sys/*' --exclude '/tmp/*' // /mnt/clone
+mount "/dev/${dst_base}2" "/mnt/clone"
+sleep 1
+rsync ${rsync_options} --delete \
+	--exclude '.gvfs' \
+	--exclude '/dev/*' \
+	--exclude '/mnt/clone/*' \
+	--exclude '/proc/*' \
+	--exclude '/run/*' \
+	--exclude '/sys/*' \
+	--exclude '/tmp/*' \
+	// /mnt/clone
 sync
 
 if [ ! -d /mnt/clone/boot ]; then
@@ -109,15 +130,18 @@ if [ ! -d /mnt/clone/boot ]; then
 	mkdir -p /mnt/clone/boot
 fi
 
-mount /dev/${dst_base}1 /mnt/clone/boot
-sync
-rsync ${rsync_options} --delete --exclude '.gvfs' --exclude 'lost\+found/*' /boot/ /mnt/clone/boot
+mount "/dev/${dst_base}1" "/mnt/clone/boot"
+sleep 1
+rsync ${rsync_options} --delete \
+	--exclude '.gvfs' \
+	--exclude 'lost\+found/*' \
+	/boot/ /mnt/clone/boot
 sync
 
 df -h
 
 cmdline=$(cat /mnt/clone/boot/cmdline.txt)
-echo -e "\n\033[38;5;211mOriginal cmdline\033[m"
+echo -e "\n\033[38;5;211mOriginal cmdline.txt\033[m"
 echo "${cmdline}"
 
 fstab=$(cat /mnt/clone/etc/fstab)
@@ -127,12 +151,12 @@ echo "${fstab}"
 for p in 1 2; do
 	src_uuid=$(lsblk -n -o PARTUUID /dev/${src_base}${p})
 	dst_uuid=$(lsblk -n -o PARTUUID /dev/${dst_base}${p})
-	#echo "src = ${src_uuid}   dst = ${dst_uuid}"
+	log "p = ${p}   src = ${src_uuid}   dst = ${dst_uuid}"
 	cmdline=$(echo "${cmdline}" | sed -e "s/${src_uuid}/${dst_uuid}/g")
 	fstab=$(echo "${fstab}" | sed -e "s/${src_uuid}/${dst_uuid}/g")
 done
 
-echo -e "\n\033[38;5;82mNew cmdline\033[m"
+echo -e "\n\033[38;5;82mNew cmdline.txt\033[m"
 echo "${cmdline}"
 echo "${cmdline}" > /mnt/clone/boot/cmdline.txt
